+++ packages/babel/src/index.ts
@@ -1,12 +1,10 @@
 import { declare } from '@babel/helper-plugin-utils';
-import type {
-  ImportSpecifier,
-  JSXAttribute,
-  ObjectProperty,
-} from '@babel/types';
-import type { Locale } from '@helven/i18n';
-import fs from 'fs-extra';
 import path from 'path';
+import { Adapter, createHelpers, getAdapter, Locale } from './utils';
+
+export type { Locale };
+
+export const defaultAdapter = getAdapter();
 
 const EXCLUDE_PATH = 'node_modules';
 
@@ -32,15 +30,18 @@ export type Options = {
     /** extracted translate component name, default to `Trans` */
     component?: string;
   };
-  /** debug mode */
-  debug?: boolean;
+  /** adapter to read and write json file */
+  adapter?: Adapter;
 };
 
 declare module '@babel/core' {
   interface PluginPass {
-    messages: Set<string | undefined>;
-    targets: Required<NonNullable<Options['extract']>>;
-    targetFunction: string;
+    cache: {
+      messages: any[];
+      targets: Required<NonNullable<Options['extract']>>;
+      targetFunction: string;
+      shouldSkip: boolean;
+    };
   }
 }
 
@@ -51,12 +52,12 @@ declare module '@babel/types' {
 }
 
 // TODO: extract with scope
-const BabelPluginI18n = declare(({ types: t }, options: Options, dirname) => {
+const BabelPluginI18n = declare((babel, options: Options, dirname) => {
   const {
     strict = true,
     langs = [],
     output = './locales',
-    debug = false,
+    adapter = defaultAdapter,
     // TODO: auto translate messages
     // translator,
   } = options;
@@ -72,73 +73,26 @@ const BabelPluginI18n = declare(({ types: t }, options: Options, dirname) => {
 
   const distination = path.resolve(dirname, output);
 
-  // import {} from 'i18n'
-  const isModule = (node: any, name: string) => {
-    return (
-      t.isImportDeclaration(node) &&
-      t.isStringLiteral(node.source, { value: name })
-    );
-  };
-
-  // import { useTranslation, Trans } from 'i18n'
-  const getImportName = (nodes: any[], name: string) => {
-    const target: ImportSpecifier = nodes.find(
-      (node) =>
-        t.isImportSpecifier(node) && t.isIdentifier(node.imported, { name }),
-    );
-    return target?.local.name ?? name;
-  };
-
-  // t()
-  const isCall = (node: any, name: string) => {
-    return t.isCallExpression(node) && t.isIdentifier(node.callee, { name });
-  };
-
-  // i18n.t()
-  const isMemberCall = (node: any, object: string, property: string) => {
-    return (
-      t.isCallExpression(node) &&
-      t.isMemberExpression(node.callee) &&
-      t.isIdentifier(node.callee.object, { name: object }) &&
-      t.isIdentifier(node.callee.property, { name: property })
-    );
-  };
-
-  // <Trans></Trans>
-  const isComponent = (node: any, name: string) => {
-    return (
-      t.isJSXElement(node) &&
-      t.isJSXIdentifier(node.openingElement.name, { name })
-    );
-  };
-
-  const getPropertyName = (nodes: any[], name: string) => {
-    const target: ObjectProperty = nodes.find(
-      (node) => t.isObjectProperty(node) && t.isIdentifier(node.key, { name }),
-    );
-    return t.isIdentifier(target?.value) ? target.value.name : undefined;
-  };
-
-  const getStringValue = (node: any) => {
-    const target = t.isJSXExpressionContainer(node) ? node.expression : node;
-    if (t.isJSXText(target) || t.isStringLiteral(target)) {
-      return target.value;
-    }
-  };
-
-  const getJSXAttrValue = (nodes: any[], name: string) => {
-    const target: JSXAttribute = nodes.find(
-      (node) =>
-        t.isJSXAttribute(node) && t.isJSXIdentifier(node.name, { name }),
-    );
-    return target?.value;
-  };
+  const {
+    t,
+    isImportModule,
+    isCall,
+    isMemberCall,
+    isComponent,
+    getImportName,
+    getPropertyName,
+    getStringValue,
+    getJSXAttrValue,
+  } = createHelpers(babel);
 
   return {
     pre() {
-      this.messages = new Set();
-      this.targets = { ...extract };
-      this.targetFunction = this.targets.function;
+      this.cache = {
+        messages: [],
+        targets: { ...extract },
+        targetFunction: extract.function,
+        shouldSkip: false,
+      };
 
       if (this.filename?.includes(EXCLUDE_PATH)) {
         this.file.path.stop();
@@ -146,12 +100,18 @@ const BabelPluginI18n = declare(({ types: t }, options: Options, dirname) => {
     },
     visitor: {
       /** stop if no use target package in strict mode */
-      Program(path, { targets }) {
+      Program(path, { cache }) {
+        const { targets } = cache;
+
         if (
           strict &&
-          !path.get('body').some(({ node }) => isModule(node, targets.package))
+          !path
+            .get('body')
+            .some(({ node }) => isImportModule(node, targets.package))
         ) {
-          path.stop();
+          // cannot use `path.skip()` directly, may cause building error
+          // why?
+          cache.shouldSkip = true;
         }
       },
       /**
@@ -162,8 +122,14 @@ const BabelPluginI18n = declare(({ types: t }, options: Options, dirname) => {
        * import { useTranslation as useT } from '@helven/i18n'
        * import { Trans as Translate } from '@helven/i18n'
        */
-      ImportDeclaration({ node }, { targets }) {
-        if (strict && isModule(node, targets.package)) {
+      ImportDeclaration({ node }, { cache }) {
+        const { targets, shouldSkip } = cache;
+
+        if (shouldSkip) {
+          return;
+        }
+
+        if (strict && isImportModule(node, targets.package)) {
           targets.hook = getImportName(node.specifiers, targets.hook);
           targets.component = getImportName(node.specifiers, targets.component);
         }
@@ -177,11 +143,11 @@ const BabelPluginI18n = declare(({ types: t }, options: Options, dirname) => {
        * }
        */
       BlockStatement: {
-        enter(path, state) {
-          path.node.targetFunction = state.targetFunction;
+        enter(path, { cache }) {
+          path.node.targetFunction = cache.targetFunction;
         },
-        exit(path, state) {
-          state.targetFunction = path.node.targetFunction;
+        exit(path, { cache }) {
+          cache.targetFunction = path.node.targetFunction;
         },
       },
       /**
@@ -191,17 +157,21 @@ const BabelPluginI18n = declare(({ types: t }, options: Options, dirname) => {
        * const { t } = useTranslate()
        * const { t: translate } = useTranslate()
        */
-      VariableDeclarator({ node }, state) {
+      VariableDeclarator({ node }, { cache }) {
+        if (cache.shouldSkip) {
+          return;
+        }
+
         if (
-          isCall(node.init, state.targets.hook) &&
+          isCall(node.init!, cache.targets.hook) &&
           t.isObjectPattern(node.id)
         ) {
           const properName = getPropertyName(
             node.id.properties,
-            state.targets.function,
+            cache.targets.function,
           );
           if (properName) {
-            state.targetFunction = properName;
+            cache.targetFunction = properName;
           }
         }
       },
@@ -212,13 +182,19 @@ const BabelPluginI18n = declare(({ types: t }, options: Options, dirname) => {
        * t('hello') // or aliasT('hello')
        * i18n.t('hello')
        */
-      CallExpression({ node }, { targets, targetFunction, messages }) {
+      CallExpression({ node }, { cache }) {
+        const { shouldSkip, targets, targetFunction, messages } = cache;
+
+        if (shouldSkip) {
+          return;
+        }
+
         if (
           isCall(node, targetFunction) ||
           isMemberCall(node, targets.instance, targets.function)
         ) {
           const value = getStringValue(node.arguments[0]);
-          messages.add(value);
+          messages.push(value);
         }
       },
       /**
@@ -230,41 +206,35 @@ const BabelPluginI18n = declare(({ types: t }, options: Options, dirname) => {
        * <Trans message="hello" />
        * <Trans message={"hello"} />
        */
-      JSXElement({ node }, { targets, messages }) {
+      JSXElement({ node }, { cache }) {
+        const { shouldSkip, targets, messages } = cache;
+
+        if (shouldSkip) {
+          return;
+        }
+
         if (isComponent(node, targets.component)) {
           const attrs = node.openingElement.attributes;
           const value =
             getStringValue(node.children) ??
             getStringValue(getJSXAttrValue(attrs, 'message'));
-          messages.add(value);
+          messages.push(value);
         }
       },
     },
     post() {
-      if (this.messages.size === 0) {
+      const { messages } = this.cache;
+
+      if (messages.length === 0) {
         return;
       }
 
       for (const lang of langs) {
         const file = path.resolve(distination, lang + '.json');
-        const locales: Locale = fs.readJSONSync(file, { throws: false }) ?? {};
-
-        let hasChanges = false;
-
-        this.messages.forEach((key) => {
-          if (key && !locales[key]) {
-            locales[key] = key;
-            hasChanges = true;
-          }
-        });
-
-        if (hasChanges) {
-          if (!debug) {
-            fs.outputJsonSync(file, locales, { spaces: 2 });
-          } else {
-            console.log('Extracted: %s.json', lang);
-            console.log(JSON.stringify(locales, null, 2));
-          }
+        const source = adapter.read(file);
+        const locales = adapter.merge(source, messages);
+        if (locales) {
+          adapter.write(file, locales);
         }
       }
     },