+++ packages/babel/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @helven/i18n-babel
 
+## 1.1.0
+
+### Minor Changes
+
+- support wepback plugin
+
 ## 1.0.5
 
 ### Patch Changes