+++ packages/webpack/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @helven/i18n-webpack
 
+## 1.1.0
+
+### Minor Changes
+
+- support wepback plugin
+
+### Patch Changes
+
+- Updated dependencies
+  - @helven/babel-plugin-i18n@1.1.0
+
 ## 1.0.5
 
 ### Patch Changes